refactor:users/prices/auth
