+++ backend/dockerfile
@@ -17,8 +17,8 @@ COPY . .
 EXPOSE 8000
 
 # 定義環境變量
-ENV MODULE_NAME="app.main"
+ENV MODULE_NAME="app.src.main"
 ENV VARIABLE_NAME="app"
 
 # 使用 uvicorn 運行 fastapi 應用
-CMD ["uvicorn", "app.main:app", "--host", "0.0.0.0", "--port", "8000"]
+CMD ["uvicorn", "app.src.main:app", "--host", "0.0.0.0", "--port", "8000"]
