--- backend/main.py
@@ -1,529 +1,0 @@
-import json
-import sentry_sdk
-from apscheduler.schedulers.background import BackgroundScheduler
-from fastapi.middleware.cors import CORSMiddleware
-import itertools
-from sqlalchemy import delete, insert, select
-from sqlalchemy.orm import Session, sessionmaker
-from typing import List, Optional
-import requests
-from fastapi import APIRouter, HTTPException, Query, Depends, status, FastAPI
-import os
-from datetime import datetime, timedelta
-from fastapi.security import OAuth2PasswordBearer, OAuth2PasswordRequestForm
-from jose import JWTError, jwt
-from passlib.context import CryptContext
-
-from pydantic import BaseModel, Field, AnyHttpUrl
-from sqlalchemy import (Column, ForeignKey, Integer, String, Table, Text,
-                        create_engine)
-from sqlalchemy.ext.declarative import declarative_base
-from sqlalchemy.orm import relationship, sessionmaker
-
-Base = declarative_base()
-
-
-user_news_association_table = Table(
-    "user_news_upvotes",
-    Base.metadata,
-    Column("user_id", Integer, ForeignKey("users.id"), primary_key=True),
-    Column(
-        "news_articles_id", Integer, ForeignKey("news_articles.id"), primary_key=True
-    ),
-)
-
-# from pydantic import BaseModel
-class Constants:
-    MAX_USERNAME_LENGTH = 50
-    MAX_PASSWORD_LENGTH = 200
-    KEY = '1892dhianiandowqd0n'
-    NEWS_LINK = "https://udn.com/api/more"
-
-
-class User(Base):
-    __tablename__ = "users"
-    id = Column(Integer, primary_key=True, autoincrement=True)
-    username = Column(String(Constants.MAX_USERNAME_LENGTH), unique=True, nullable=False)
-    hashed_password = Column(String(Constants.MAX_PASSWORD_LENGTH), nullable=False)
-    upvoted_news = relationship(
-        "NewsArticle",
-        secondary=user_news_association_table,
-        back_populates="upvoted_by_users",
-    )
-
-
-class NewsArticle(Base):
-    __tablename__ = "news_articles"
-    id = Column(Integer, primary_key=True, autoincrement=True)
-    url = Column(String, unique=True, nullable=False)
-    title = Column(String, nullable=False)
-    time = Column(String, nullable=False)
-    content = Column(Text, nullable=False)
-    summary = Column(Text, nullable=False)
-    reason = Column(Text, nullable=False)
-    upvoted_by_users = relationship(
-        "User", secondary=user_news_association_table, back_populates="upvoted_news"
-    )
-
-
-engine = create_engine("sqlite:///news_database.db", echo=True)
-
-Base.metadata.create_all(engine)
-
-Session = sessionmaker(bind=engine)
-
-sentry_sdk.init(
-    dsn="https://4001ffe917ccb261aa0e0c34026dc343@o4505702629834752.ingest.us.sentry.io/4507694792704000",
-    traces_sample_rate=1.0,
-    profiles_sample_rate=1.0,
-)
-
-app = FastAPI()
-Scheduler = BackgroundScheduler()
-SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
-
-app.add_middleware(
-    CORSMiddleware,  # noqa
-    allow_origins=["http://localhost:8080"],
-    allow_credentials=True,
-    allow_methods=["*"],
-    allow_headers=["*"],
-)
-
-import os
-from openai import OpenAI
-
-
-# def generate_summary(content):
-#     m = [
-#         {
-#             "role": "system",
-#             "content": "你是一個新聞摘要生成機器人，請統整新聞中提及的影響及主要原因 (影響、原因各50個字，請以json格式回答 {'影響': '...', '原因': '...'})",
-#         },
-#         {"role": "user", "content": f"{content}"},
-#     ]
-#
-#     completion = OpenAI(api_key="xxx").chat.completions.create(
-#         model="gpt-3.5-turbo",
-#         messages=m,
-#     )
-#     return completion.choices[0].message.content
-
-#
-# def extract_search_keywords(content):
-#     m = [
-#         {
-#             "role": "system",
-#             "content": "你是一個關鍵字提取機器人，用戶將會輸入一段文字，表示其希望看見的新聞內容，請提取出用戶希望看見的關鍵字，請截取最重要的關鍵字即可，避免出現「新聞」、「資訊」等混淆搜尋引擎的字詞。(僅須回答關鍵字，若有多個關鍵字，請以空格分隔)",
-#         },
-#         {"role": "user", "content": f"{content}"},
-#     ]
-#
-#     completion = OpenAI(api_key="xxx").chat.completions.create(
-#         model="gpt-3.5-turbo",
-#         messages=m,
-#     )
-#     return completion.choices[0].message.content
-
-
-from urllib.parse import quote
-import requests
-from bs4 import BeautifulSoup
-from sqlalchemy.orm import Session
-
-
-def add_news_article(news_data):
-    """
-    add new to db
-    :param news_data: news info
-    :return:
-    """
-    session = Session()
-    session.add(NewsArticle(
-        url=news_data["url"],
-        title=news_data["title"],
-        time=news_data["time"],
-        content=" ".join(news_data["content"]),  # 將內容list轉換為字串
-        summary=news_data["summary"],
-        reason=news_data["reason"],
-    ))
-    session.commit()
-    session.close()
-
-
-def fetch_news_articles_by_keyword(search_term, is_initial=False):
-    """
-    get new by keyword
-
-    :param search_term: keyword
-    :param is_initial: bool indicate whether  this is the initial fetch
-    :return:
-    """
-    all_news_data = []
-    # iterate pages to get more news data, not actually get all news data
-    if is_initial:
-        news = []
-        for page in range(1, 10):
-            pageinfo = {
-                "page": page,
-                "id": f"search:{quote(search_term)}",
-                "channelId": 2,
-                "type": "searchword",
-            }
-            response = requests.get(Constants.NEWS_LINK, params=pageinfo)
-            news.append(response.json()["lists"])
-
-        for result in news:
-            all_news_data.append(result)
-    else:
-        pageinfo = {
-            "page": 1,
-            "id": f"search:{quote(search_term)}",
-            "channelId": 2,
-            "type": "searchword",
-        }
-        response = requests.get(Constants.NEWS_LINK, params=pageinfo)
-
-        all_news_data = response.json()["lists"]
-    return all_news_data
-
-def get_new_info(is_initial=False):
-    """
-    get new info
-
-    :param is_initial:
-    :return:
-    """
-    news_data = fetch_news_articles_by_keyword("價格", is_initial=is_initial)
-    for news in news_data:
-        title = news["title"]
-        summary_prompt = [
-            {
-                "role": "system",
-                "content": "你是一個關聯度評估機器人，請評估新聞標題是否與「民生用品的價格變化」相關，並給予'high'、'medium'、'low'評價。(僅需回答'high'、'medium'、'low'三個詞之一)",
-            },
-            {"role": "user", "content": f"{title}"},
-        ]
-        summary_completion = OpenAI(api_key="xxx").chat.completions.create(
-            model="gpt-3.5-turbo",
-            messages=GPTinfo,
-        )
-        relevance = summary_completion.choices[0].message.content
-        if relevance == "high":
-            response = requests.get(news["titleLink"])
-            soup = BeautifulSoup(response.text, "html.parser")
-            # 標題
-            title = soup.find("h1", class_="article-content__title").text
-            time = soup.find("time", class_="article-content__time").text
-            # 定位到包含文章內容的 <section>
-            content_section = soup.find("section", class_="article-content__editor")
-
-            paragraphs = [
-                p.text
-                for p in content_section.find_all("p")
-                if p.text.strip() != "" and "?" not in p.text
-            ]
-            detailed_news =  {
-                "url": news["titleLink"],
-                "title": title,
-                "time": time,
-                "content": paragraphs,
-            }
-            GPTinfo = [
-                {
-                    "role": "system",
-                    "content": "你是一個新聞摘要生成機器人，請統整新聞中提及的影響及主要原因 (影響、原因各50個字，請以json格式回答 {'影響': '...', '原因': '...'})",
-                },
-                {"role": "user", "content": " ".join(detailed_news["content"])},
-            ]
-
-            completion = OpenAI(api_key="xxx").chat.completions.create(
-                model="gpt-3.5-turbo",
-                messages=GPTinfo,
-            )
-            result = completion.choices[0].message.content
-            result = json.loads(result)
-            detailed_news["summary"] = result["影響"]
-            detailed_news["reason"] = result["原因"]
-            add_news_article(detailed_news)
-
-
-@app.on_event("startup")
-def start_scheduler():
-    database = SessionLocal()
-    if database.query(NewsArticle).count() == 0:
-        # should change into simple factory pattern
-        get_new_info()
-    database.close()
-    Scheduler.add_job(get_new_info, "interval", minutes=100)
-    Scheduler.start()
-
-
-@app.on_event("shutdown")
-def shutdown_scheduler():
-    Scheduler.shutdown()
-
-
-pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
-oauth2_scheme = OAuth2PasswordBearer(tokenUrl="/api/v1/users/login")
-
-
-def session_opener():
-    session = Session(bind=engine)
-    try:
-        yield session
-    finally:
-        session.close()
-
-
-
-def verify(password, hashed_password):
-    return pwd_context.verify(password, hashed_password)
-
-
-def check_user_password_is_correct(database, name, pwd):
-    user = database.query(User).filter(User.username == name).first()
-    if not verify(pwd, user.hashed_password):
-        return False
-    return user
-
-
-def authenticate_user_token(
-    token = Depends(oauth2_scheme),
-    database = Depends(session_opener)
-):
-    payload = jwt.decode(token, Constants.KEY, algorithms=["HS256"])
-    return database.query(User).filter(User.username == payload.get("sub")).first()
-
-
-def create_access_token(data, expires_delta=None):
-    """create access token"""
-    to_encode = data.copy()
-    if expires_delta:
-        expire = datetime.utcnow() + expires_delta
-    else:
-        expire = datetime.utcnow() + timedelta(minutes=15)
-    to_encode.update({"exp": expire})
-    print(to_encode)
-    encoded_json_webtoken = jwt.encode(to_encode, Constants.KEY, algorithm="HS256")
-    return encoded_json_webtoken
-
-
-@app.post("/api/v1/users/login")
-async def login_for_access_token(
-        form_data: OAuth2PasswordRequestForm = Depends(), database: Session = Depends(session_opener)
-):
-    """login"""
-    user = check_user_password_is_correct(database, form_data.username, form_data.password)
-    access_token = create_access_token(
-        data={"sub": str(user.username)}, expires_delta=timedelta(minutes=30)
-    )
-    return {"access_token": access_token, "token_type": "bearer"}
-
-class UserAuthSchema(BaseModel):
-    username: str
-    password: str
-@app.post("/api/v1/users/register")
-def create_user(user: UserAuthSchema, database: Session = Depends(session_opener)):
-    """create user"""
-    hashed_password = pwd_context.hash(user.password)
-    database_user = User(username=user.username, hashed_password=hashed_password)
-    database.add(database_user)
-    database.commit()
-    database.refresh(database_user)
-    return database_user
-
-
-@app.get("/api/v1/users/me")
-def read_users_me(user=Depends(authenticate_user_token)):
-    return {"username": user.username}
-
-
-_id_counter = itertools.count(start=1000000)
-
-
-def get_article_upvote_details(article_id, uid, database):
-    counter = (
-        database.query(user_news_association_table)
-        .filter_by(news_articles_id=article_id)
-        .count()
-    )
-    voted = False
-    if uid:
-        voted = (
-                database.query(user_news_association_table)
-                .filter_by(news_articles_id=article_id, user_id=uid)
-                .first()
-                is not None
-        )
-    return counter, voted
-
-
-@app.get("/api/v1/news/news")
-def read_news(database=Depends(session_opener)):
-    """
-    read new
-
-    :param db:
-    :return:
-    """
-    news = database.query(NewsArticle).order_by(NewsArticle.time.desc()).all()
-    result = []
-    for new in news:
-        upvotes, upvoted = get_article_upvote_details(new.id, None, database)
-        result.append(
-            {**new.__dict__, "upvotes": upvotes, "is_upvoted": upvoted}
-        )
-    return result
-
-
-@app.get("/api/v1/news/user_news")
-def read_user_news(
-        database=Depends(session_opener),
-        user=Depends(authenticate_user_token)
-):
-    """
-    read user new
-
-    :param db:
-    :param u:
-    :return:
-    """
-    news = database.query(NewsArticle).order_by(NewsArticle.time.desc()).all()
-    result = []
-    for article in news:
-        upvotes, upvoted = get_article_upvote_details(article.id, user.id, database)
-        result.append(
-            {
-                **article.__dict__,
-                "upvotes": upvotes,
-                "is_upvoted": upvoted,
-            }
-        )
-    return result
-
-class PromptRequest(BaseModel):
-    prompt: str
-
-@app.post("/api/v1/news/search_news")
-async def search_news(request: PromptRequest):
-    prompt = request.prompt
-    news_list = []
-    keyword_prompt  = [
-        {
-            "role": "system",
-            "content": "你是一個關鍵字提取機器人，用戶將會輸入一段文字，表示其希望看見的新聞內容，請提取出用戶希望看見的關鍵字，請截取最重要的關鍵字即可，避免出現「新聞」、「資訊」等混淆搜尋引擎的字詞。(僅須回答關鍵字，若有多個關鍵字，請以空格分隔)",
-        },
-        {"role": "user", "content": f"{prompt}"},
-    ]
-
-    completion = OpenAI(api_key="xxx").chat.completions.create(
-        model="gpt-3.5-turbo",
-        messages=keyword_prompt ,
-    )
-    keywords = completion.choices[0].message.content
-    # should change into simple factory pattern
-    news_items = fetch_news_articles_by_keyword(keywords, is_initial=False)
-    for news in news_items:
-        try:
-            response = requests.get(news["titleLink"])
-            soup = BeautifulSoup(response.text, "html.parser")
-            # 標題
-            title = soup.find("h1", class_="article-content__title").text
-            time = soup.find("time", class_="article-content__time").text
-            # 定位到包含文章?容的 <section>
-            content_section = soup.find("section", class_="article-content__editor")
-
-            paragraphs = [
-                p.text
-                for p in content_section.find_all("p")
-                if p.text.strip() != "" and "?" not in p.text
-            ]
-            detailed_news = {
-                "url": news["titleLink"],
-                "title": title,
-                "time": time,
-                "content": paragraphs,
-            }
-            detailed_news["content"] = " ".join(detailed_news["content"])
-            detailed_news["id"] = next(_id_counter)
-            news_list.append(detailed_news)
-        except Exception as error_message:
-            print(error_message)
-    return sorted(news_list, key=lambda x: x["time"], reverse=True)
-
-class NewsSumaryRequestSchema(BaseModel):
-    content: str
-
-@app.post("/api/v1/news/news_summary")
-async def news_summary(
-        payload: NewsSumaryRequestSchema, user=Depends(authenticate_user_token)
-):
-    response = {}
-    summary_prompt = [
-        {
-            "role": "system",
-            "content": "你是一個新聞摘要生成機器人，請統整新聞中提及的影響及主要原因 (影響、原因各50個字，請以json格式回答 {'影響': '...', '原因': '...'})",
-        },
-        {"role": "user", "content": f"{payload.content}"},
-    ]
-
-    completion = OpenAI(api_key="xxx").chat.completions.create(
-        model="gpt-3.5-turbo",
-        messages=summary_prompt ,
-    )
-    result = completion.choices[0].message.content
-    if result:
-        result = json.loads(result)
-        response["summary"] = result["影響"]
-        response["reason"] = result["原因"]
-    return response
-
-
-@app.post("/api/v1/news/{id}/upvote")
-def upvote_article(
-        id,
-        database=Depends(session_opener),
-        user=Depends(authenticate_user_token),
-):
-    message = toggle_upvote(id, user.id, database)
-    return {"message": message}
-
-
-def toggle_upvote(article_id, user_id, database):
-    existing_upvote = database.execute(
-        select(user_news_association_table).where(
-            user_news_association_table.c.news_articles_id == article_id,
-            user_news_association_table.c.user_id == user_id,
-        )
-    ).scalar()
-
-    if existing_upvote:
-        delete_stmt = delete(user_news_association_table).where(
-            user_news_association_table.c.news_articles_id == article_id,
-            user_news_association_table.c.user_id == user_id,
-        )
-        database.execute(delete_stmt)
-        database.commit()
-        return "Upvote removed"
-    else:
-        insert_stmt = insert(user_news_association_table).values(
-            news_articles_id=article_id, user_id=user_id
-        )
-        database.execute(insert_stmt)
-        database.commit()
-        return "Article upvoted"
-
-
-def news_exists(id2, database: Session):
-    return database.query(NewsArticle).filter_by(id=id2).first() is not None
-
-
-@app.get("/api/v1/prices/necessities-price")
-def get_necessities_prices(
-        category=Query(None), commodity=Query(None)
-):
-    return requests.get(
-        "https://opendata.ey.gov.tw/api/ConsumerProtection/NecessitiesPrice",
-        params={"CategoryName": category, "Name": commodity},
-    ).json()
