+++ backend/src/auth/constant.py
@@ -1,0 +1,2 @@
+MAX_USERNAME_LENGTH = 50
+MAX_PASSWORD_LENGTH = 200
