+++ backend/src/database.py
@@ -1,0 +1,18 @@
+from sqlalchemy import (Column, ForeignKey, Integer, String, Table, Text,
+                        create_engine)
+from sqlalchemy.ext.declarative import declarative_base
+from sqlalchemy.orm import relationship, sessionmaker
+Base = declarative_base()
+
+engine = create_engine("sqlite:///news_database.db", echo=True)
+
+Base.metadata.create_all(engine)
+
+Session = sessionmaker(bind=engine)
+
+def session_opener():
+    session = Session()
+    try:
+        yield session
+    finally:
+        session.close()
