+++ backend/src/models.py
@@ -1,0 +1,12 @@
+from sqlalchemy import (Column, ForeignKey, Integer, String, Table, Text,
+                        create_engine)
+from .database import Base
+
+user_news_association_table = Table(
+    "user_news_upvotes",
+    Base.metadata,
+    Column("user_id", Integer, ForeignKey("users.id"), primary_key=True),
+    Column(
+        "news_articles_id", Integer, ForeignKey("news_articles.id"), primary_key=True
+    ),
+)
