+++ backend/src/news/constant.py
@@ -1,0 +1,1 @@
+NEWS_LINK = "https://udn.com/api/more"
