+++ backend/src/news/models.py
@@ -1,0 +1,17 @@
+from ..database import Base
+from ..models import user_news_association_table
+from sqlalchemy import (Column, ForeignKey, Integer, String, Table, Text,
+                        create_engine)
+from sqlalchemy.orm import relationship
+class NewsArticle(Base):
+    __tablename__ = "news_articles"
+    id = Column(Integer, primary_key=True, autoincrement=True)
+    url = Column(String, unique=True, nullable=False)
+    title = Column(String, nullable=False)
+    time = Column(String, nullable=False)
+    content = Column(Text, nullable=False)
+    summary = Column(Text, nullable=False)
+    reason = Column(Text, nullable=False)
+    upvoted_by_users = relationship(
+        "User", secondary=user_news_association_table, back_populates="upvoted_news"
+    )
