+++ backend/src/news/schemas.py
@@ -1,0 +1,5 @@
+from pydantic import BaseModel
+class PromptRequest(BaseModel):
+    prompt: str
+class NewsSumaryRequestSchema(BaseModel):
+    content: str
