+++ backend/src/users/schemas.py
@@ -1,0 +1,4 @@
+from pydantic import BaseModel
+class UserAuthSchema(BaseModel):
+    username: str
+    password: str
